Complete reconciliation lifecycle seams
