+++ internal/reconcile/manager.go
@@ -9,23 +9,11 @@ import (
 	"sync"
 	"time"
 
-	"github.com/hashicorp/nomad/api"
-	"github.com/hashicorp/nomad/jobspec2"
-
 	"github.com/brianmichel/nomad-compass/internal/manifest"
 	"github.com/brianmichel/nomad-compass/internal/nomadclient"
 	bundleplan "github.com/brianmichel/nomad-compass/internal/plan"
 	"github.com/brianmichel/nomad-compass/internal/repo"
 	"github.com/brianmichel/nomad-compass/internal/storage"
-)
-
-const (
-	compassMetaRepoURL      = "nomad-compass/repo-url"
-	compassMetaRepoName     = "nomad-compass/repo-name"
-	compassMetaJobFile      = "nomad-compass/job-file"
-	compassMetaCommit       = "nomad-compass/commit"
-	compassMetaCommitAuthor = "nomad-compass/commit-author"
-	compassMetaCommitTitle  = "nomad-compass/commit-title"
 )
 
 // Manager coordinates reconciliation cycles for onboarded repositories.
@@ -194,66 +182,18 @@ func (m *Manager) PlanRepo(ctx context.Context, repoID int64) (*bundleplan.Repor
 			case "job":
 				status, err := m.nomad.JobStatus(ctx, resource.Name)
 				return status != nil && status.Exists, err
-			case "acl_policy":
-				policy, err := resourceLookup.ObserveACLPolicy(ctx, resource.Name)
-				return policy != nil, err
-			case "volume":
-				spec, err := manifest.CompileVolume(resource)
-				if err != nil {
-					return false, err
-				}
-				if spec.Type == "csi" && resourceClient != nil {
-					id := spec.CSI.ID
-					if id == "" {
-						id = resource.Name
-					}
-					volume, err := resourceClient.ObserveCSIVolume(ctx, id, effectiveNamespace(spec.CSI.Namespace))
-					return volume != nil, err
-				}
-				if spec.Type == "host" {
-					volume, err := resourceLookup.FindHostVolume(ctx, spec.Host.Name, spec.Host.Namespace)
-					return volume != nil, err
-				}
-				volume, err := resourceLookup.FindCSIVolume(ctx, spec.CSI.Name, spec.CSI.Namespace)
-				return volume != nil, err
 			default:
-				return false, nil
+				adapter, ok := adapterFor(resource.Kind)
+				if !ok {
+					return false, nil
+				}
+				return adapter.Lookup(ctx, resourceLookup, resource)
 			}
 		}
 	}
 	result, err := bundleplan.CompareTrackedWithLookup(ctx, snapshot.Bundle, tracked, func(ctx context.Context, resource manifest.Resource, tracked storage.ManagedResource) (bundleplan.Observation, error) {
 		if resource.Kind == "job" {
-			candidateID := tracked.NomadID.String
-			if candidateID == "" {
-				source, sourceErr := manifest.NativeJobSource(resource)
-				if sourceErr != nil {
-					return bundleplan.Observation{}, sourceErr
-				}
-				job, _, parseErr := parseJob(resource.SourcePath, source)
-				if parseErr != nil {
-					return bundleplan.Observation{}, parseErr
-				}
-				candidateID = jobID(job)
-			}
-			status, err := m.nomad.JobStatus(ctx, candidateID)
-			if err != nil || status == nil || !status.Exists {
-				return bundleplan.Observation{Present: status != nil && status.Exists}, err
-			}
-			source, err := manifest.NativeJobSource(resource)
-			if err != nil {
-				return bundleplan.Observation{}, err
-			}
-			job, _, err := parseJob(resource.SourcePath, source)
-			if err != nil {
-				return bundleplan.Observation{}, err
-			}
-			annotateJob(job, repoRecord, repo.JobFile{Path: resource.Address, Content: source}, snapshot, false)
-			job.ID = &tracked.NomadID.String
-			jobPlan, err := m.nomad.PlanJob(ctx, job)
-			if err != nil {
-				return bundleplan.Observation{}, err
-			}
-			return bundleplan.Observation{Present: true, Matches: !jobPlanHasChanges(jobPlan)}, nil
+			return m.observeBundleJob(ctx, resource, tracked)
 		}
 		if resourceClient == nil {
 			return bundleplan.Observation{}, errors.New("Nomad client does not support bundle resources")
@@ -325,30 +265,11 @@ func (m *Manager) adoptBundleResource(ctx context.Context, repoID int64, snapsho
 
 	var nomadID, namespace, subtype string
 	if resource.Kind == "job" {
-		status, err := m.nomad.JobStatus(ctx, resource.Name)
+		result, err := m.adoptBundleJob(ctx, resource)
 		if err != nil {
 			return err
 		}
-		if status == nil || !status.Exists {
-			return fmt.Errorf("Nomad job %q does not exist", resource.Name)
-		}
-		source, err := manifest.NativeJobSource(resource)
-		if err != nil {
-			return err
-		}
-		job, _, err := parseJob(resource.SourcePath, source)
-		if err != nil {
-			return err
-		}
-		job.ID = &resource.Name
-		jobPlan, err := m.nomad.PlanJob(ctx, job)
-		if err != nil {
-			return err
-		}
-		if jobPlanHasChanges(jobPlan) {
-			return fmt.Errorf("Nomad job %q does not match desired bundle resource", resource.Name)
-		}
-		nomadID = resource.Name
+		nomadID, namespace, subtype = result.NomadID, result.Namespace, result.Subtype
 	} else {
 		lookup, ok := m.nomad.(nomadclient.ResourceLookup)
 		if !ok {
@@ -365,19 +286,6 @@ func (m *Manager) adoptBundleResource(ctx context.Context, repoID int64, snapsho
 		nomadID, namespace, subtype = result.NomadID, result.Namespace, result.Subtype
 	}
 	return m.upsertManagedResource(ctx, repoID, snapshot, resource, nomadID, namespace, manifest.SpecHash(resource), manifest.ManifestHash(resource), "applied", "", subtype)
-}
-
-func (m *Manager) applyJob(ctx context.Context, repoRecord *storage.Repository, jobFile repo.JobFile, snapshot *repo.Snapshot, job *api.Job, submission *api.JobSubmission) (string, error) {
-	if job == nil || submission == nil {
-		return "", errors.New("job and submission are required")
-	}
-
-	annotateJob(job, repoRecord, jobFile, snapshot, true)
-
-	if err := m.nomad.RegisterJob(ctx, job, submission); err != nil {
-		return "", err
-	}
-	return jobID(job), nil
 }
 
 // ListProtectedResources returns resources that Compass retained after they
@@ -569,149 +477,6 @@ func (m *Manager) unscheduleJobs(ctx context.Context, repoID int64) error {
 			return err
 		}
 	}
-	return nil
-}
-
-func jobID(job *api.Job) string {
-	if job == nil {
-		return ""
-	}
-	if job.ID != nil && *job.ID != "" {
-		return *job.ID
-	}
-	if job.Name != nil {
-		return *job.Name
-	}
-	return ""
-}
-
-func parseJob(path string, contents []byte) (*api.Job, *api.JobSubmission, error) {
-	cfg := &jobspec2.ParseConfig{Body: contents, Path: path, Strict: true}
-	job, err := jobspec2.ParseWithConfig(cfg)
-	if err != nil {
-		return nil, nil, err
-	}
-	if job.Meta == nil {
-		job.Meta = map[string]string{}
-	}
-	submission := &api.JobSubmission{
-		Source: string(contents),
-		Format: "hcl2",
-	}
-	return job, submission, nil
-}
-
-func (m *Manager) ensureJobs(ctx context.Context, repoRecord *storage.Repository, snapshot *repo.Snapshot, commitChanged bool) error {
-	jobFiles, err := snapshotJobFiles(snapshot)
-	if err != nil {
-		return err
-	}
-
-	repoFiles, err := m.files.ListByRepo(ctx, repoRecord.ID)
-	if err != nil {
-		return err
-	}
-
-	fileIndex := make(map[string]storage.RepoFile, len(repoFiles))
-	for _, file := range repoFiles {
-		fileIndex[file.Path] = file
-	}
-
-	seen := make(map[string]struct{}, len(jobFiles))
-
-	for _, jobFile := range jobFiles {
-		seen[jobFile.Path] = struct{}{}
-		existing, tracked := fileIndex[jobFile.Path]
-		job, submission, err := parseJob(jobFile.Path, jobFile.Content)
-		if err != nil {
-			m.logger.Error("job parse failed", "repo", repoRecord.Name, "file", jobFile.Path, "error", err)
-			continue
-		}
-
-		needApply := !tracked
-		var trackedJobID string
-		if tracked && existing.JobID.Valid {
-			trackedJobID = existing.JobID.String
-		}
-
-		if tracked && !needApply {
-			if trackedJobID == "" {
-				needApply = true
-			} else {
-				status, err := m.nomad.JobStatus(ctx, trackedJobID)
-				if err != nil {
-					m.logger.Warn("job status check failed", "repo", repoRecord.Name, "job_id", trackedJobID, "file", jobFile.Path, "error", err)
-					if commitChanged {
-						needApply = true
-					} else {
-						continue
-					}
-				}
-				if status == nil || !status.Exists {
-					needApply = true
-				}
-			}
-		}
-
-		if tracked && !needApply {
-			annotateJob(job, repoRecord, jobFile, snapshot, false)
-			plan, err := m.nomad.PlanJob(ctx, job)
-			if err != nil {
-				m.logger.Warn("job plan failed", "repo", repoRecord.Name, "job_id", trackedJobID, "file", jobFile.Path, "error", err)
-				needApply = true
-			} else if !jobPlanHasChanges(plan) {
-				if commitChanged {
-					if err := m.files.UpsertWithDeleteMode(ctx, repoRecord.ID, jobFile.Path, snapshot.CommitHash, trackedJobID, jobFile.DeleteMode); err != nil {
-						return err
-					}
-				}
-				continue
-			} else {
-				needApply = true
-			}
-		}
-
-		if !needApply {
-			continue
-		}
-
-		jobID, err := m.applyJob(ctx, repoRecord, jobFile, snapshot, job, submission)
-		if err != nil {
-			m.logger.Error("job apply failed", "repo", repoRecord.Name, "file", jobFile.Path, "error", err)
-			continue
-		}
-		if err := m.files.UpsertWithDeleteMode(ctx, repoRecord.ID, jobFile.Path, snapshot.CommitHash, jobID, jobFile.DeleteMode); err != nil {
-			return err
-		}
-	}
-
-	for path, file := range fileIndex {
-		if _, ok := seen[path]; ok {
-			continue
-		}
-		// Job file no longer exists in the repo. Unschedule and drop tracking metadata.
-		if file.DeleteMode.Valid && file.DeleteMode.String == string(manifest.DeleteModeProtect) {
-			if m.logger != nil {
-				m.logger.Warn("job removal protected", "repo", repoRecord.Name, "file", path, "job_id", file.JobID.String)
-			}
-			continue
-		}
-		if file.JobID.Valid && file.JobID.String != "" {
-			if err := m.nomad.DeregisterJob(ctx, file.JobID.String, true); err != nil {
-				if m.logger != nil {
-					m.logger.Error("job deregister failed", "repo", repoRecord.Name, "job_id", file.JobID.String, "file", path, "error", err)
-				}
-				continue
-			}
-		}
-		if err := m.files.Delete(ctx, repoRecord.ID, path); err != nil {
-			return err
-		}
-		if m.logger != nil {
-			m.logger.Info("job removed", "repo", repoRecord.Name, "file", path, "job_id", file.JobID.String)
-		}
-	}
-
 	return nil
 }
 
@@ -960,8 +725,7 @@ func encodedDependencies(resource manifest.Resource) string {
 
 func validateBundleResources(resources []manifest.Resource) error {
 	for _, resource := range resources {
-		switch resource.Kind {
-		case "job":
+		if resource.Kind == "job" {
 			source, err := manifest.NativeJobSource(resource)
 			if err != nil {
 				return err
@@ -969,78 +733,13 @@ func validateBundleResources(resources []manifest.Resource) error {
 			if _, _, err := parseJob(resource.SourcePath, source); err != nil {
 				return fmt.Errorf("validate bundle job %q: %w", resource.Address, err)
 			}
-		case "volume", "acl_policy":
-			adapter, ok := adapterFor(resource.Kind)
-			if !ok {
-				return fmt.Errorf("bundle resource %q is not supported yet", resource.Address)
-			}
-			if err := adapter.Validate(resource); err != nil {
-				return err
-			}
-		default:
-			return fmt.Errorf("bundle resource %q is not supported yet", resource.Address)
-		}
-	}
-	return nil
-}
-
-func (m *Manager) ensureBundleJobs(ctx context.Context, repoRecord *storage.Repository, snapshot *repo.Snapshot, ordered []manifest.Resource, tracked map[string]storage.ManagedResource) error {
-	for _, resource := range ordered {
-		if resource.Kind != "job" {
 			continue
 		}
-
-		source, err := manifest.NativeJobSource(resource)
-		if err != nil {
-			return err
+		adapter, ok := adapterFor(resource.Kind)
+		if !ok {
+			return fmt.Errorf("bundle resource %q is not supported yet", resource.Address)
 		}
-		job, submission, err := parseJob(resource.Address, source)
-		if err != nil {
-			return fmt.Errorf("parse bundle job %q: %w", resource.Address, err)
-		}
-
-		hash := manifest.SpecHash(resource)
-		manifestHash := manifest.ManifestHash(resource)
-		existing := tracked[resource.Address]
-		jobFile := repo.JobFile{Path: resource.Address, Content: source, DeleteMode: string(resource.DeleteMode)}
-		trackedJobID := existing.NomadID.String
-
-		if existing.Address == "" {
-			candidateID := jobID(job)
-			status, statusErr := m.nomad.JobStatus(ctx, candidateID)
-			if statusErr != nil {
-				return fmt.Errorf("check ownership for bundle job %q: %w", resource.Address, statusErr)
-			}
-			if status != nil && status.Exists {
-				return fmt.Errorf("bundle job %q collides with unmanaged Nomad job %q", resource.Address, candidateID)
-			}
-		}
-
-		if existing.Address != "" && trackedJobID != "" {
-			status, statusErr := m.nomad.JobStatus(ctx, trackedJobID)
-			if statusErr != nil {
-				if m.logger != nil {
-					m.logger.Warn("bundle job status check failed", "repo", repoRecord.Name, "job", resource.Address, "error", statusErr)
-				}
-			} else if status != nil && status.Exists {
-				job.ID = &trackedJobID
-				annotateJob(job, repoRecord, jobFile, snapshot, false)
-				plan, planErr := m.nomad.PlanJob(ctx, job)
-				if planErr == nil && !jobPlanHasChanges(plan) {
-					if err := m.upsertManagedResource(ctx, repoRecord.ID, snapshot, resource, trackedJobID, "", hash, manifestHash, "applied", "", ""); err != nil {
-						return err
-					}
-					continue
-				}
-			}
-		}
-
-		appliedID, applyErr := m.applyJob(ctx, repoRecord, jobFile, snapshot, job, submission)
-		if applyErr != nil {
-			_ = m.upsertManagedResource(ctx, repoRecord.ID, snapshot, resource, trackedJobID, "", hash, manifestHash, "failed", applyErr.Error(), "")
-			return fmt.Errorf("apply bundle job %q: %w", resource.Address, applyErr)
-		}
-		if err := m.upsertManagedResource(ctx, repoRecord.ID, snapshot, resource, appliedID, "", hash, manifestHash, "applied", "", ""); err != nil {
+		if err := adapter.Validate(resource); err != nil {
 			return err
 		}
 	}
@@ -1076,16 +775,6 @@ func (m *Manager) managedResources(ctx context.Context, repoID int64) (map[strin
 	return result, nil
 }
 
-func (m *Manager) replaceManagedVolume(ctx context.Context, client nomadclient.ResourceClient, resource manifest.Resource, tracked storage.ManagedResource) error {
-	if resource.DeleteMode != manifest.DeleteModeAllow {
-		return fmt.Errorf("volume changes are protected; use a new resource address or set delete = %q to allow replacement", manifest.DeleteModeAllow)
-	}
-	if err := deleteManagedResource(ctx, client, tracked); err != nil {
-		return fmt.Errorf("replace volume %q: delete existing volume: %w", resource.Address, err)
-	}
-	return nil
-}
-
 func (m *Manager) ensureManagedResource(ctx context.Context, repoID int64, snapshot *repo.Snapshot, client nomadclient.ResourceClient, resource manifest.Resource, tracked storage.ManagedResource) error {
 	adapter, ok := adapterFor(resource.Kind)
 	if !ok {
@@ -1113,8 +802,8 @@ func (m *Manager) ensureManagedResource(ctx context.Context, repoID int64, snaps
 		if observeErr != nil {
 			return observeErr
 		}
-		if resource.Kind == "volume" && observation.Present && (!observation.Matches || tracked.ContentHash.Valid && tracked.ContentHash.String != hash) {
-			err = m.replaceManagedVolume(ctx, client, resource, tracked)
+		if observation.Present && (!observation.Matches || tracked.ContentHash.Valid && tracked.ContentHash.String != hash) {
+			err = adapter.Replace(ctx, client, resource, tracked)
 		}
 	}
 	if err != nil {
@@ -1146,7 +835,7 @@ func (m *Manager) ensureManagedResource(ctx context.Context, repoID int64, snaps
 	}
 	var result managedResourceResult
 	if err == nil {
-		result, err = adapter.Apply(ctx, client, resource)
+		result, err = adapter.Apply(ctx, client, resource, tracked)
 	}
 	if err != nil {
 		return m.recordManagedResourceFailure(ctx, repoID, snapshot, resource, tracked, failureHash, manifestHash, err, "apply")
@@ -1175,116 +864,12 @@ func (m *Manager) recordManagedResourceFailure(ctx context.Context, repoID int64
 }
 
 func deleteManagedResource(ctx context.Context, client nomadclient.ResourceClient, resource storage.ManagedResource) error {
-	switch resource.Kind {
-	case "job":
+	if resource.Kind == "job" {
 		return client.DeregisterJob(ctx, resource.NomadID.String, true)
-	case "volume":
-		if resource.Subtype.Valid && resource.Subtype.String == "csi" {
-			return client.DeleteCSIVolume(ctx, resource.NomadID.String, resource.Namespace.String, false)
-		}
-		return client.DeleteHostVolume(ctx, resource.NomadID.String, resource.Namespace.String, false)
-	case "acl_policy":
-		return client.DeleteACLPolicy(ctx, resource.NomadID.String)
-	default:
+	}
+	adapter, ok := adapterFor(resource.Kind)
+	if !ok {
 		return fmt.Errorf("resource kind %q cannot be deleted", resource.Kind)
 	}
-}
-
-func snapshotJobFiles(snapshot *repo.Snapshot) ([]repo.JobFile, error) {
-	if snapshot == nil {
-		return nil, errors.New("snapshot is required")
-	}
-	if snapshot.Bundle != nil {
-		return nil, errors.New("bundle jobs use managed-resource tracking")
-	}
-	return snapshot.JobFiles, nil
-}
-
-func jobPlanHasChanges(resp *api.JobPlanResponse) bool {
-	if resp == nil {
-		return true
-	}
-	return jobDiffHasChanges(resp.Diff)
-}
-
-func jobDiffHasChanges(diff *api.JobDiff) bool {
-	if diff == nil {
-		return false
-	}
-	if fieldDiffsHaveChanges(diff.Fields) || len(diff.Objects) > 0 {
-		return true
-	}
-	for _, tg := range diff.TaskGroups {
-		if taskGroupDiffHasChanges(tg) {
-			return true
-		}
-	}
-	return false
-}
-
-func taskGroupDiffHasChanges(diff *api.TaskGroupDiff) bool {
-	if diff == nil {
-		return false
-	}
-	if fieldDiffsHaveChanges(diff.Fields) || len(diff.Objects) > 0 {
-		return true
-	}
-	for _, task := range diff.Tasks {
-		if taskDiffHasChanges(task) {
-			return true
-		}
-	}
-	return false
-}
-
-func taskDiffHasChanges(diff *api.TaskDiff) bool {
-	if diff == nil {
-		return false
-	}
-	if fieldDiffsHaveChanges(diff.Fields) || len(diff.Objects) > 0 {
-		return true
-	}
-	return false
-}
-
-func fieldDiffsHaveChanges(fields []*api.FieldDiff) bool {
-	for _, field := range fields {
-		if field == nil || isCompassCommitMetadataField(field.Name) {
-			continue
-		}
-		return true
-	}
-	return false
-}
-
-func isCompassCommitMetadataField(name string) bool {
-	switch name {
-	case nomadMetaFieldName(compassMetaCommit),
-		nomadMetaFieldName(compassMetaCommitAuthor),
-		nomadMetaFieldName(compassMetaCommitTitle):
-		return true
-	default:
-		return false
-	}
-}
-
-func nomadMetaFieldName(key string) string {
-	return "Meta[" + key + "]"
-}
-
-func annotateJob(job *api.Job, repoRecord *storage.Repository, jobFile repo.JobFile, snapshot *repo.Snapshot, includeCommitMetadata bool) {
-	if job == nil {
-		return
-	}
-	if job.Meta == nil {
-		job.Meta = map[string]string{}
-	}
-	job.Meta[compassMetaRepoURL] = repoRecord.RepoURL
-	job.Meta[compassMetaRepoName] = repoRecord.Name
-	job.Meta[compassMetaJobFile] = jobFile.Path
-	if includeCommitMetadata && snapshot != nil {
-		job.Meta[compassMetaCommit] = snapshot.CommitHash
-		job.Meta[compassMetaCommitAuthor] = snapshot.CommitAuthor
-		job.Meta[compassMetaCommitTitle] = snapshot.CommitTitle
-	}
+	return adapter.Delete(ctx, client, resource)
 }
